requirements.txt
# Gemini API相关依赖
google-genai>=0.8.0

# 图片处理
Pillow>=10.0.0

# 数据验证
pydantic>=2.0.0

# 基础依赖
pathlib
typing-extensions
